Move User code for billing and billing webhooks to Enterprise-Exclusive
